requirements.txt
certifi==2021.10.8
chardet==3.0.4
click==7.1.2
fastapi==0.62.0
FastAPI-SQLAlchemy==0.2.1
fastapi-utils==0.2.1
greenlet==1.1.2
gunicorn==20.0.4
h11==0.12.0
httptools==0.1.1
idna==2.10
importlib-metadata==4.8.2
psycopg2-binary==2.8.6
python-dotenv==0.19.1
pydantic==1.8.2
PyJWT==1.4.2
python-json-logger==0.1.11
python-multipart==0.0.5
redis==3.5.3
requests==2.24.0
six==1.16.0
SQLAlchemy==1.4.27
starlette==0.13.6
typing_extensions==4.0.0
urllib3==1.25.11
uvicorn==0.12.3
uvloop==0.14.0
zipp==3.6.0
opentelemetry-exporter-jaeger==1.6.2
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-fastapi==0.26b1
opentelemetry-instrumentation-logging==0.26b1
opentelemetry-instrumentation-requests==0.26b1
minio==7.1.2

